first draft of getConfig/ and setSpannerConfig/
